requirements.txt
amqp==5.2.0
appdirs==1.4.4
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.2.0
bleach==6.1.0
celery==5.3.6
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==0.7
cron-descriptor==1.4.0
cryptography==41.0.7
defusedxml==0.8.0rc2
distlib==0.3.7
Django==4.2.8
django-allauth==0.59.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-filter==23.5
django-microsoft-auth==3.0.1
django-phone-field==1.8.1
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
filelock==3.12.2
h11==0.14.0
idna==3.6
importlib-metadata==4.8.3
kombu==5.3.4
lxml==4.9.3
Markdown==3.3.7
martor==1.6.28
numpy==1.26.0
oauthlib==3.2.2
outcome==1.2.0
packaging==23.1
pandas==2.1.1
Pillow==10.1.0
platformdirs==3.9.1
prompt-toolkit==3.0.43
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
PySocks==1.7.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
selenium==4.13.0
six==1.16.0
flower==2.0.1
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
sweetify==2.3.1
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.0.4
vine==5.1.0
virtualenv==20.24.2
wcwidth==0.2.12
gunicorn==20.0.4
dj-static==0.0.6